src/test/java/algorithms/countingvalleys/SolutionTest.java
package algorithms.countingvalleys;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class SolutionTest implements ArgumentsProvider {

  @Override
  public Stream<? extends Arguments> provideArguments(ExtensionContext extensionContext) {
    return Stream.of(
      Arguments.of(1, "UDDDUDUU"),
      Arguments.of(2, "DUDDDUUDUU"),
      Arguments.of(6, "UDUDDUUUDUDUDUUDUUDDDDDUDUDDDDUUDDUDDUUUUDUUDUDDDDUDUDUUUDDDUUUDUDDUUDDDUUDDUDDDUDUUDUUDUUDUDDDUUUUU")
    );
  }

  @ParameterizedTest
  @ArgumentsSource(SolutionTest.class)
  void countingValleys(int expected, String input) {
    assertEquals(expected, Solution.countingValleys(input));
  }
}
